mysite/requirements.txt
djangorestframework
Django==2.1.4
pytz==2017.3